feat: add explicit error messages
