+++ src/managers/mod.rs
@@ -1,3 +1,4 @@
+pub mod error;
 pub mod issuer;
 pub mod relying_party;
 pub mod status_provider;
